% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_randomization.R
\name{balance_randomization}
\alias{balance_randomization}
\title{Randomize a dataset, checking for balance with Lawley-Hotelling statistic}
\usage{
balance_randomization(
  data,
  block_vars,
  cluster_vars = NULL,
  balance_vars = NULL,
  n_treatment = 2,
  attempts = 10,
  condition_names = NULL,
  treat_probs = 0.5,
  print_status = FALSE,
  ...
)
}
\arguments{
\item{data}{a dataframe or tibble to be randomized}

\item{block_vars}{a vector of quoted column names for blocking variables}

\item{cluster_vars}{an optional vector of quoted column names to cluster on}

\item{balance_vars}{a vector of quoted column names for checking balance}

\item{n_treatment}{the number of experimental groups to randomize into}

\item{attempts}{the number of randomizations to conduct. the most balanced of these randomizations will be returned}

\item{condition_names}{a vector of quoted names of the experimental groups}

\item{treat_probs}{a vector of proportion to be placed into each group. If only two experimental groups, you may specifcy the proportion for one group and the remaining will be in the other group.}

\item{print_status}{boolean, will print status if set to TRUE}
}
\value{
a vector of treatment assignments
}
\description{
This function is a wrapper around the randomizr functions. It runs a
randomization several times, selecting the one that is most balanced based
on the specified balance variables. The best balance is measured using the
Lawley-Hotelling statistic, which was added with the help of Brian Osserman.
}
\details{
The L-W statistic is a measure of balance which is independent of affine
transformations of covariates (in particular, it does not depend on units).
However, it cannot distinguish more important covariates from less important
ones, so it is advisable not to include unnecessary covariates among the
balance variables.

The L-W statistic has some strange behavior when some covariates are highly
correlated, and in the extreme of 100\% correlation it will fail due to being
unable to invert a matrix.
}
\examples{
iris$assignments <- balance_randomization(data = iris,
block_vars = "Species",
balance_vars = c("Sepal.Length", "Sepal.Width"))

#'\dontrun{library(nycflights13)
randomized_data <- flights[complete.cases(flights), ] \%>\%
filter(carrier \%in\% c("AA", "DL")) \%>\%
head(10000) \%>\%
balance_randomization(n_treatment = 3,
                        block_vars = c("carrier"),
                        cluster_vars = c("carrier", "tailnum"),
                        balance_vars = c("air_time", "distance"),
                        treat_probs = c(.2, .2, .6))}
}
